requirements.txt
absl-py==1.3.0
accelerate==0.15.0
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.5.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==39.0.1
diffusers==0.9.0
Django==4.1.3
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.8.0
flatbuffers==22.12.6
frozenlist==1.3.3
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==5.1.0
keras==2.11.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.4
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.2
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
pandas-stubs==1.5.2.221124
Pillow==9.3.0
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soundfile==0.11.0
sqlparse==0.4.3
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
tokenizers==0.13.2
torch==1.13.0
tqdm==4.64.1
transformers==4.24.0
types-pytz==2022.6.0.1
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.2
zipp==3.11.0
##pip freeze > requirements.txt